- GAME IS NOW A SINGLETON!!!!! - All ships must be of ShipBase - Normal ships are now BasicShips
