+++ src/main/java/controllers/AttackGameAction.java
@@ -5,13 +5,9 @@ import cs361.battleships.models.Game;
 
 public class AttackGameAction {
 
-    @JsonProperty private Game game;
     @JsonProperty private int x;
     @JsonProperty private char y;
 
-    public Game getGame() {
-        return game;
-    }
 
     public int getActionRow() {
         return x;
