--- src/main/java/cs361/battleships/models/Ship.java
@@ -1,152 +1,0 @@
-package cs361.battleships.models;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.google.common.collect.Sets;
-
-import java.util.*;
-
-public class Ship {
-
-	@JsonProperty private String kind;
-	@JsonProperty private List<Square> occupiedSquares;
-	@JsonProperty private int size;
-	@JsonIgnore   private int cqPos;
-    @JsonProperty private Square CQuarters;
-    @JsonProperty private boolean isArmored;
-
-
-	public Ship() {
-		occupiedSquares = new ArrayList<>();
-	}
-	
-	public Ship(String kind) {
-		this();
-		this.kind = kind;
-		switch(kind) {
-			case "MINESWEEPER":
-				size = 2;
-				cqPos = 0;
-				isArmored = false;
-				break;
-			case "DESTROYER":
-				size = 3;
-				cqPos = 1;
-				isArmored = true;
-				break;
-			case "BATTLESHIP":
-				size = 4;
-				cqPos = 2;
-				isArmored = true;
-				break;
-		}
-	}
-
-	public List<Square> getOccupiedSquares() {
-		return occupiedSquares;
-	}
-
-	public void place(char col, int row, boolean isVertical) {
-	    Square temp;
-		for (int i=0; i<size; i++) {
-            if (isVertical) {
-                temp = new Square(row + i, col);
-
-            } else {
-                temp = new Square(row, (char) (col + i));
-            }
-            if(i == this.cqPos) {
-                CQuarters = temp;
-                CQuarters.setCQ(true);
-            }
-            this.occupiedSquares.add(temp);
-        }
-
-	}
-
-	public boolean overlaps(Ship other) {
-		Set<Square> thisSquares = Set.copyOf(getOccupiedSquares());
-		Set<Square> otherSquares = Set.copyOf(other.getOccupiedSquares());
-		Sets.SetView<Square> intersection = Sets.intersection(thisSquares, otherSquares);
-		return intersection.size() != 0;
-	}
-
-	public boolean isAtLocation(Square location) {
-		return getOccupiedSquares().stream().anyMatch(s -> s.equals(location));
-	}
-
-	public String getKind() {
-		return kind;
-	}
-
-	public Result attack(int x, char y) {
-		var attackedLocation = new Square(x, y);
-		var square = getOccupiedSquares().stream().filter(s -> s.equals(attackedLocation)).findFirst();
-		if (!square.isPresent()) {
-			return new Result(attackedLocation);
-		}
-		var attackedSquare = square.get();
-        var result = new Result(attackedLocation);
-        if (attackedSquare.getCQ())
-        {
-            if (!this.isArmored && !this.isSunk())
-            {
-                getOccupiedSquares().forEach( elem -> elem.hit());
-                result.setShip(this);
-                result.setResult(AttackStatus.SUNK);
-                return result;
-            }
-            else if (this.isSunk())
-            {
-                result.setResult(AttackStatus.INVALID);
-                return result;
-            }
-            else
-            {
-                this.isArmored = false;
-                result.setShip(this);
-                result.setResult(AttackStatus.MISS);
-                return result;
-            }
-        }
-        if (attackedSquare.isHit() && !attackedSquare.getCQ()) {
-            result.setResult(AttackStatus.INVALID);
-            return result;
-        }
-		attackedSquare.hit();
-		result.setShip(this);
-		if (isSunk()) {
-			result.setResult(AttackStatus.SUNK);
-		} else {
-			result.setResult(AttackStatus.HIT);
-		}
-		return result;
-	}
-
-	@JsonIgnore
-	public boolean isSunk() {
-		return getOccupiedSquares().stream().allMatch(s -> s.isHit());
-	}
-
-	@Override
-	public boolean equals(Object other) {
-		if (!(other instanceof Ship)) {
-			return false;
-		}
-		var otherShip = (Ship) other;
-
-		return this.kind.equals(otherShip.kind)
-				&& this.size == otherShip.size
-				&& this.occupiedSquares.equals(otherShip.occupiedSquares);
-	}
-
-	@Override
-	public int hashCode() {
-		return 33 * kind.hashCode() + 23 * size + 17 * occupiedSquares.hashCode();
-	}
-
-	@Override
-	public String toString() {
-		return kind + occupiedSquares.toString();
-	}
-}
