+++ src/main/java/cs361/battleships/models/Ships/ShipUtility.java
@@ -1,0 +1,12 @@
+package cs361.battleships.models.Ships;
+
+public class ShipUtility {
+
+    public static ShipBase createShip(String kind) {
+        if(kind == "Submarine") {
+            return null;
+        }
+        return new BasicShip(kind);
+    }
+
+}
